completed the remove-invalid-matches functionality
