+++ main/helper.py
@@ -20,7 +20,20 @@ def get_invalid_matches():
 
     return(invalid_matches)
 
+
 def remove_invalid_matches():
     invalid_matches = get_invalid_matches()
     for match in invalid_matches:
         match.delete()
+
+
+def get_match_details(match):
+    return {
+        "id": match.id,
+        "date_and_time": match.date_and_time,
+        "player1": match.player1.name,
+        "player2": match.player2.name,
+        "player1_score": match.player1_score,
+        "player2_score": match.player2_score,
+        "field": match.field.name
+    }
